requirements.txt
# Основные зависимости
Jinja2>=3.0
Pillow>=9.0

# Для работы с LLM API (выберите нужные)
google-generativeai>=0.3.0  # Для Gemini API
groq>=0.4.0  # Для Groq API
# openai>=1.0.0  # Для OpenAI API (опционально)

# Для парсинга PDF (выберите одну)
PyPDF2>=3.0.0  # Простой парсер PDF
# pdfplumber>=0.10.0  # Альтернатива (лучше работает с таблицами)
